requirements.txt
Pillow>=10
pygame>=2.5.0
